adapt init script for proper shutdown behaviour
The LSB header now demands the nutdown daemon to be stopped
only after the root filesystem is mounted readonly which should
happen just before total system halt.

To make this possible, a symlink in /run/sendsigs.omit.d is also
necessary for nutdown to not get SIGKILLed at shutdown time. This
may be Debian/Ubuntu specific.